Refactor forc-test to expose interpreter
